[ADD] Nueva visualizacion a tab detalles

+++ lib/app/modules/calculadora/calculadora_controller.dart
@@ -12,7 +12,7 @@ class CalculadoraController extends GetxController{
 
   RxInt consumo  = 0.obs;
   RxDouble costo = 0.0.obs;
-  RxList<int> listConsumo = List<int>().obs;
+  RxList<double> listConsumo = List<double>().obs;
   RxList<double> listPrecio = List<double>().obs;
 
   RxBool expanded = false.obs;
@@ -46,10 +46,10 @@ class CalculadoraController extends GetxController{
     consumo.value = (lectura2.value - lectura1.value).abs(); //? OJO: abs() retorna el valor absoluto (|x|)
     listPrecio.clear();
     listConsumo.clear();
-    Map<String, dynamic> resultado = calcCosto(consumo.value);
+    Map<String, dynamic> resultado = calcCosto(consumo.value.toDouble());
     
     costo.value = resultado["costo"]; // es de tipo RxDouble
-    List<int> lc = resultado["listaConsumo"];
+    List<double> lc = resultado["listaConsumo"];
     lc.forEach((e) { listConsumo.add(e);});
     List<double> lp = resultado["listaPrecio"];
     lp.forEach((e) { listPrecio.add(e);});
